Add disjoint set data structure
This disjoint set uses a tree under the hood.
The only thing that a node in this tree needs is
a pointer to its parent. To determine if a node is
in the set, all that needs to be done is a walk up
the tree until the root is found. Each set corresponds
to a different root.