Simplify disjoint set find
Find was also in charge of flattening the tree,
but its better design practice to keep concerns
separated. The only reason we need flattening is
for unions, so we just flatten when we are doing
that instead of every time there's a find.

Also, since we're only flattening when we have to
finding is more efficient. If find was called on an
already flattened tree then it would have had to
make two passes through it for no reason.